requirements.txt
cookies==2.2.1
mock==1.0.1
Pillow==2.8.2
python-forecastio==1.3.3
requests==2.7.0
responses==0.4.0
six==1.9.0
wheel==0.24.0
